TODO_App/src/main/java/github/MSzleag/logic/TaskGroupService.java
package github.MSzleag.logic;


import github.MSzleag.model.TaskGroup;
import github.MSzleag.model.TaskGroupRepository;
import github.MSzleag.model.TaskRepository;
import github.MSzleag.model.projection.GroupReadModel;
import github.MSzleag.model.projection.GroupWriteModel;

import java.util.*;
import java.util.stream.Collectors;


class TaskGroupService {
    private TaskGroupRepository repository;
    private TaskRepository taskRepository;

    public TaskGroupService(final TaskGroupRepository repository, final TaskRepository taskRepository) {
        this.repository = repository;
        this.taskRepository = taskRepository;
    }

    public GroupReadModel createGroup(final GroupWriteModel source){
        TaskGroup result = repository.save(source.toGroup());
        return new GroupReadModel(result);
    }

    public List<GroupReadModel> readAll(){
        return repository.findAll()
                .stream()
                .map(GroupReadModel::new)
                .collect(Collectors.toList());
    }

    public void toggleGroup(int groupId) {
        if (taskRepository.existsByDoneIsFalseAndGroup_Id(groupId)){
            throw new IllegalStateException("Group has undone tasks. Done all the tasks first");
        }
        TaskGroup result = repository.findById(groupId)
                .orElseThrow(() -> new IllegalArgumentException("TaskGroup with given id not found"));
        result.setDone(!result.isDone());
        repository.save(result);
    }
}
